refactor(database): cleanup redundant code from associations
