--- modules/database/src/adapters/sequelize-adapter/sql-adapter/SQLSchema.ts
@@ -1,121 +1,0 @@
-import { Sequelize, Transaction } from 'sequelize';
-import { ConduitDatabaseSchema, SingleDocQuery } from '../../../interfaces';
-import ConduitGrpcSdk, { Indexable } from '@conduitplatform/grpc-sdk';
-import { SQLAdapter } from './index';
-import { SequelizeSchema } from '../SequelizeSchema';
-import { getTransactionAndParsedQuery } from '../utils';
-
-const incrementDbQueries = () =>
-  ConduitGrpcSdk.Metrics?.increment('database_queries_total');
-
-export class SQLSchema extends SequelizeSchema {
-  constructor(
-    sequelize: Sequelize,
-    schema: Indexable,
-    originalSchema: ConduitDatabaseSchema,
-    adapter: SQLAdapter,
-    extractedRelations: { [key: string]: SequelizeSchema | SequelizeSchema[] },
-  ) {
-    super(sequelize, schema, originalSchema, adapter, extractedRelations);
-  }
-
-  async findByIdAndUpdate(
-    id: string,
-    query: SingleDocQuery,
-    populate?: string[],
-    transaction?: Transaction,
-  ): Promise<Indexable> {
-    const { t, parsedQuery, transactionProvided } = await getTransactionAndParsedQuery(
-      transaction,
-      query,
-      this.sequelize,
-    );
-    try {
-      const parentDoc = await this.model.findByPk(id, {
-        nest: true,
-        include: this.constructRelationInclusion(populate),
-        transaction: t,
-      });
-      if (parsedQuery.hasOwnProperty('$inc')) {
-        const inc = parsedQuery['$inc'];
-        for (const key in inc) {
-          if (!inc.hasOwnProperty(key)) continue;
-          if (this.extractedRelations[key]) {
-            throw new Error(`Cannot increment relation: ${key}`);
-          }
-          if (inc[key] > 0) {
-            parsedQuery[key] = Sequelize.literal(`${key} + ${inc[key]}`);
-          } else {
-            parsedQuery[key] = Sequelize.literal(`${key} - ${Math.abs(inc[key])}`);
-          }
-        }
-        delete parsedQuery['$inc'];
-      }
-
-      if (parsedQuery.hasOwnProperty('$push')) {
-        const push = parsedQuery['$push'];
-        for (const key in push) {
-          if (this.extractedRelations[key]) {
-            throw new Error(`Cannot increment relation: ${key}`);
-          }
-          if (this.extractedRelations[key]) {
-            if (!Array.isArray(this.extractedRelations[key])) {
-              throw new Error(`Cannot push in non-array field: ${key}`);
-            }
-            let modelName = key.charAt(0).toUpperCase() + key.slice(1);
-            if (push[key]['$each']) {
-              if (!modelName.endsWith('s')) {
-                modelName = modelName + 's';
-              }
-              parentDoc[`add${modelName}`](push[key]['$each'], parentDoc._id);
-            } else {
-              const actualRel = key.charAt(0).toUpperCase() + key.slice(1);
-              parentDoc[`add${actualRel}Id`](push[key], parentDoc._id);
-            }
-            continue;
-          }
-          if (push[key]['$each']) {
-            parentDoc[key] = [...parentDoc[key], ...push[key]['$each']];
-          } else {
-            parentDoc[key] = [...parentDoc[key], push[key]];
-          }
-        }
-        await parentDoc.save();
-        delete parsedQuery['$push'];
-      }
-
-      parsedQuery.updatedAt = new Date();
-      incrementDbQueries();
-      const relationObjects = this.extractRelationsModification(parsedQuery);
-      await this.model.update({ ...parsedQuery }, { where: { _id: id }, transaction: t });
-      incrementDbQueries();
-
-      const data = await this.model
-        .findByPk(id, {
-          nest: true,
-          include: this.constructRelationInclusion(populate),
-          transaction: t,
-        })
-        .then(doc => this.createWithPopulation(doc, relationObjects, t!))
-        .then(() => {
-          if (!transactionProvided) {
-            return t!.commit();
-          }
-          return;
-        })
-        .then(() => {
-          return this.model.findByPk(id, {
-            nest: true,
-            include: this.constructRelationInclusion(populate),
-          });
-        })
-        .then(doc => (doc ? doc.toJSON() : doc));
-      return data;
-    } catch (e) {
-      if (!transactionProvided) {
-        await t!.rollback();
-      }
-      throw e;
-    }
-  }
-}
